Remove some gcc-4.3.4 warnings (Cygwin)
The latest gcc with the latest Cygwin complains when passing a char to
the toupper, tolower, isspace, isalnum, isprint, isdigit or isalpha
functions/macros. These functions are defined to take an integer. This
patch adds cast to int as needed to remove the warnings. After this
there are still two warnings related to signed/unsigned comparison in
yy_get_next_buffer() (part of flex).

+++ vpi/sys_display.c
@@ -1907,7 +1907,7 @@ static PLI_INT32 sys_severity_calltf(PLI_BYTE8*name)
 
       /* convert name to upper and drop $ to get severity string */
       sstr = strdup(name) + 1;
-      for (t=sstr; *t; t+=1) *t = toupper(*t);
+      for (t=sstr; *t; t+=1) *t = toupper((int)*t);
 
       scope = vpi_handle(vpiScope, callh);
       assert(scope);
